Fix the update of an existing playlist
The result of PlaylistTrack.load_from_server() is a generator.

When the playlist is already known, we need to compare the list of
tracks from the server and in the index, which means we need to iterate
twice on the result of load_from_server().

If it's a generator, it will not work, we thus need to materialize the
result in a list.